pyproject.toml
[project]
name = "cng"
version = "1.0.4"

dependencies = ["minio",
                "ibis-framework[duckdb]",
                "duckdb",
                "streamlit",
                "pydeck",
		"boto3",
		"botocore",
		"tippecanoe",
                "huggingface_hub"]

[project.optional-dependencies]
tests = ["pytest", "pytest-cov"]

[tool.pytest.ini_options]
testpaths = ["tests"]
